+++ src/core/protocol/messages.test.ts
@@ -103,6 +103,8 @@ describe("protocol/messages", () => {
           return msg.type;
         case "setContentWidth":
           return msg.type;
+        case "setTypography":
+          return msg.type;
         case "revealLine":
           return msg.type;
         case "copied":
@@ -139,6 +141,8 @@ describe("protocol/messages", () => {
         case "jumpToSource":
           return msg.type;
         case "exportReview":
+          return msg.type;
+        case "toggleTaskCheckbox":
           return msg.type;
         default:
           return assertNever(msg);